Move assimp debug to mesh importer
